refactor(cli): split init/wizard god-module steps.rs into cohesive siblings
steps.rs was ~3830 lines, 196 symbols, the most depended-on file in the
crate. Reorganize it into a steps/ directory whose mod.rs is a curated
index over job-focused siblings:

- types.rs       step-state enums + shared CheckResult/StepState types
- dispatch.rs    titles/footers, state init, render/handle_key/validate fan-out
- indexing.rs    language toggles + directory tree
- oracle.rs      SCIP tool availability toggle + probe
- embedding.rs   model picker, remote mode/backend/server-model, tuning fields
- integration.rs version check + git/Claude hook toggles
- hooks.rs       hook-conflict choices + chained-hook renderer
- tests.rs       the inline test module, moved verbatim

Pure move/reorganize: no behavior change. mod.rs curates the same
pub(crate) surface external callers already used, so every call site keeps
its steps::* path unchanged. No re-export shims.
